pyproject.toml
[project]
name = "omnisound"
version = "1.0.0"
description = "Universal motor audio system — turn any motor into a speaker"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "OmniSound Contributors"}
]
keywords = [
    "audio",
    "motor",
    "servo",
    "esp32",
    "arduino",
    "music",
    "dsp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: System :: Hardware",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "librosa>=0.10.1",
    "pyserial>=3.5",
    "bleak>=0.21.1",
    "zeroconf>=0.131.0",
    "yt-dlp>=2024.1.0",
    "mido>=1.3.0",
    "python-rtmidi>=1.5.5",
]

[project.urls]
Homepage = "https://github.com/omnisound-project/omnisound"
Documentation = "https://omnisound-project.github.io/omnisound/docs"
Repository = "https://github.com/omnisound-project/omnisound.git"
Issues = "https://github.com/omnisound-project/omnisound/issues"
Changelog = "https://github.com/omnisound-project/omnisound/blob/main/CHANGELOG.md"

[project.scripts]
omnisound = "omnisound.omnisound:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["core*", "plugins*"]

[tool.ruff]
target-version = "py39"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "node_modules",
    "dist",
    "build",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501", "SIM102", "SIM116", "B005", "E402", "UP006", "UP035", "SIM105", "B904", "SIM910", "I001"]

[tool.ruff.lint.isort]
known-first-party = ["core", "plugins"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
exclude = '''
/(
    \.git
  | __pycache__
  | venv
  | \.venv
  | node_modules
  | dist
  | build
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
